feat: Write data into Excel sheet
prepare writing to excel

Issue #7
